requirements.txt
requests==2.28.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
selenium==4.12.0
retrying~=1.3.4
psutil~=6.0.0